move addLineNumbersToDiff method to utils
