+++ packages/processors/dify/processor.ts
@@ -1,9 +1,9 @@
 import process from 'node:process';
 import type { ReviewConfig } from '../base/types.ts';
-import { formatFileSummaryTable } from '../base/utils/formatting.ts';
+import { addLineNumbersToDiff, formatFileSummaryTable } from '../base/utils/formatting.ts';
 import { mergeOverallSummaries } from '../base/utils/summary.ts';
 import type { IFileChange, IPullRequestInfo, IPullRequestProcessedResult, IReviewComment, OverallSummary, SummarizeResult } from './deps.ts';
-import { BaseProcessor, OverallSummarySchema, type ReviewComment, ReviewCommentSchema, ReviewResponseSchema, SummaryResponseSchema } from './deps.ts';
+import { BaseProcessor, OverallSummarySchema, type ReviewComment, ReviewResponseSchema, SummaryResponseSchema } from './deps.ts';
 import { runWorkflow, uploadFile } from './internal/mod.ts';
 
 // Internal configuration type for DifyProcessor
@@ -306,7 +306,7 @@ export class DifyProcessor extends BaseProcessor {
             title: prInfo.title,
             description: prInfo.body || '',
             filePath: file.path,
-            patch: this.addLineNumbersToDiff(file.patch),
+            patch: addLineNumbersToDiff(file.patch),
             instructions: this.getInstructionsForFile(file.path, config),
             aspects: {
               transfer_method: 'local_file',
